+++ lib/shadcn_ui.ex
@@ -18,6 +18,7 @@ defmodule ShadcnUI do
     ShadcnUI.Components.Foundation.Button,
     ShadcnUI.Components.Foundation.Card,
     ShadcnUI.Components.Foundation.Skeleton,
+    ShadcnUI.Components.Forms.Checkbox,
     ShadcnUI.Components.Forms.ErrorSummary,
     ShadcnUI.Components.Forms.Field,
     ShadcnUI.Components.Forms.FieldErrors,